move the csv files package into the local package.
