+++ src/main/kotlin/data/local/csvfile/StateCsvImpl.kt
@@ -1,66 +1,65 @@
-package data.csvfile
+package org.example.data.local.csvfile
 
-import kotlinx.datetime.LocalDateTime
 import org.example.data.DataProvider
-import org.example.entity.ProjectEntity
+import org.example.entity.StateEntity
 import org.example.utils.PlanMateException
 import java.io.File
 import java.io.IOException
 import java.util.*
 
-class ProjectCsvImpl(
+class StateCsvImpl(
     fileName: String
-) : DataProvider<ProjectEntity> {
+) : DataProvider<StateEntity> {
 
     private val file: File = File(fileName)
 
-    override suspend fun add(item: ProjectEntity) {
+    override suspend fun add(item: StateEntity) {
         try {
             val items = loadFromCsv().toMutableList()
             items.add(item)
             saveToCsv(items)
         } catch (e: Exception) {
-            throw PlanMateException.FileWriteException("Error adding project: ${e.message}")
+            throw PlanMateException.FileWriteException("Error adding state: ${e.message}")
         }
     }
 
-    override suspend fun get(): List<ProjectEntity> = loadFromCsv()
+    override suspend fun get(): List<StateEntity> = loadFromCsv()
 
-    override suspend fun getById(id: UUID): ProjectEntity? = loadFromCsv().find { it.id == id }
+    override suspend fun getById(id: UUID): StateEntity? = loadFromCsv().find { it.id == id }
 
-    override suspend fun update(item: ProjectEntity) {
+    override suspend fun update(item: StateEntity) {
         val items = loadFromCsv().toMutableList()
         val index = items.indexOfFirst { it.id == item.id }
 
         if (index == -1) {
-            throw PlanMateException.ItemNotFoundException("Project with ID ${item.id} not found.")
+            throw PlanMateException.ItemNotFoundException("State with ID ${item.id} not found.")
         }
 
         items[index] = item
         try {
             saveToCsv(items)
         } catch (e: Exception) {
-            throw PlanMateException.FileWriteException("Error updating project: ${e.message}")
+            throw PlanMateException.FileWriteException("Error updating state: ${e.message}")
         }
     }
 
     override suspend fun delete(id: UUID) {
         val items = loadFromCsv().toMutableList()
-        val projectToDelete = items.find { it.id == id }
+        val stateToDelete = items.find { it.id == id }
 
-        if (projectToDelete == null) {
-            throw PlanMateException.ItemNotFoundException("Project with ID $id not found.")
+        if (stateToDelete == null) {
+            throw PlanMateException.ItemNotFoundException("State with ID $id not found.")
         }
 
-        items.remove(projectToDelete)
+        items.remove(stateToDelete)
         try {
             saveToCsv(items)
         } catch (e: Exception) {
-            throw PlanMateException.FileWriteException("Error deleting project: ${e.message}")
+            throw PlanMateException.FileWriteException("Error deleting state: ${e.message}")
         }
     }
 
-    private fun loadFromCsv(): List<ProjectEntity> {
+    private fun loadFromCsv(): List<StateEntity> {
         ensureFileExists()
         return readAndParseFile()
     }
@@ -74,15 +73,13 @@ class ProjectCsvImpl(
         }
     }
 
-    private fun readAndParseFile(): List<ProjectEntity> {
+    private fun readAndParseFile(): List<StateEntity> {
         return file.readLines()
             .filter { it.isNotBlank() }
             .map { fromCSVLine(it) }
     }
 
-
-
-    private fun saveToCsv(data: List<ProjectEntity>) {
+    private fun saveToCsv(data: List<StateEntity>) {
         try {
             val content = data.joinToString("\n") { toCSVLine(it) }
             file.writeText(content)
@@ -90,21 +87,20 @@ class ProjectCsvImpl(
             throw PlanMateException.FileWriteException("Error writing to file '${file.name}': ${e.message}")
         }
     }
-
-    private fun fromCSVLine(line: String): ProjectEntity {
+    private fun fromCSVLine(line: String): StateEntity {
         try {
             val parts = line.split(",")
-            return ProjectEntity(
+            return StateEntity(
                 id = UUID.fromString(parts[0]),
                 name = parts[1],
-                createdByAdminId = UUID.fromString(parts[2]),
-                createdAt = LocalDateTime.parse(parts[3])
+                projectId = UUID.fromString(parts[2])
             )
         } catch (e: Exception) {
             throw PlanMateException.InvalidFormatException("Malformed CSV line: $line. ${e.message}")
         }
     }
-    private fun toCSVLine(entity: ProjectEntity): String {
-        return "${entity.id},${entity.name},${entity.createdByAdminId},${entity.createdAt}"
+
+    private fun toCSVLine(entity: StateEntity): String {
+        return "${entity.id},${entity.name},${entity.projectId}"
     }
 }
