+++ src/test/kotlin/data/csvfile/UserCsvImplTest.kt
@@ -2,6 +2,7 @@ package data.csvfile
 
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.test.runTest
+import org.example.data.local.csvfile.UserCsvImpl
 import org.example.entity.UserEntity
 import org.example.entity.UserType
 import org.example.utils.PlanMateException
